Time: 0 ms (100.00%) | Memory: 7.9 MB (36.49%) - LeetSync

+++ 2050-count-good-numbers/count-good-numbers.cpp
@@ -1,17 +1,18 @@
 class Solution {
 public:
-    long long MOD = 1e9 + 7;
+    int MOD = 1e9 + 7;
 
     long long power(long long x , long long n){
+
         long long ans = 1;
 
         while(n > 0){
 
             if(n % 2 == 1){
-                ans  = (ans*x)% MOD;
+                ans = (ans * x)%MOD;
             }
 
-            x = (x*x) % MOD;
+            x = (x*x)%MOD;
             n /= 2;
         }
 
@@ -23,7 +24,7 @@ public:
         long long oddPos = n/2;
 
         long long evenWays = power(5, evenPos);
-        long long oddWays = power(4 , oddPos);
+        long long oddWays = power(4, oddPos);
 
         return (evenWays * oddWays) % MOD;
     }
